# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Intermediária TCE/SC para GPT",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://SEU-DOMINIO-AQUI"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tcesc/localizar": {
      "get": {
        "operationId": "localizarTcesc",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado da localização documental"
          }
        }
      }
    },
    "/tcesc/processo/{numero}": {
      "get": {
        "operationId": "consultarProcessoTcesc",
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta orientada de processo"
          }
        }
      }
    },
    "/tcesc/prejulgado/{numero}": {
      "get": {
        "operationId": "consultarPrejulgadoTcesc",
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta orientada de prejulgado"
          }
        }
      }
    },
    "/tcesc/municipios": {
      "get": {
        "operationId": "listarMunicipiosTcesc",
        "responses": {
          "200": {
            "description": "Municípios"
          }
        }
      }
    },
    "/tcesc/unidades-gestoras": {
      "get": {
        "operationId": "listarUnidadesGestorasTcesc",
        "responses": {
          "200": {
            "description": "Unidades gestoras"
          }
        }
      }
    }
  }
}